feat(database): schema extension support (mongo adapter)
This is still a WIP.
Creating a schema through createSchemaFromAdapter() now creates or
updates an extension for non-owners.
If called with no custom fields, the caller's extension gets deleted or
won't be created in case it doesn't already exist.
Schemas can be queried with extension fields merged into base fields
using getSchemaStitched() and getSchemasStitched().

TODO:
- Support deleting schema extensions through deleteSchema
- Implement for Sequelize
- Make getSchemaStitched() and getSchemasStitched() abstract
- Use getSchemaStitched() and getSchemasStitched() in CMS
- createAt/updateAt Date format
- Hook into CMS handlers